First pass at new `DiffFileRenderer` class
This is a foundation to base API in a somewhat similar way to diff2HTML.

This encapsulates some of the functionality that i was manually putting
together into main.

It also starts to productize the file based API a bit. Obviously next
steps would be a Diff class for full diffs, that could or could not
optionally render metadata and stuff.