Tehcnically shit is kinda working...
We got chunk based line virtualization that's sorta working. We still
get gnarly lockups when rendering c++ code though.  This is only the
first round of optimization, in that we are still doing lots of
re-rendering and still rendering in larget code chunks, which isn't
ideal...

I need to figure out how to cache results of a file and then don't
re-parse things we don't need to when scrolling back over a file...

It could be good to try and maybe keep a file around as you scroll
through it.

We definitely don't want to pay the shiki tax if possible

One downside of this version, we still get very large blanking when
jumping around, which does suck... the only way i can fix that is with
fixed positioning and request animation frame, which... well sucks for
safari...