Improve pool singleton API
Originally it was all based on an async API around being initialized,
instead I've abstracted that away to make it an implementation detail.
Also components and the like will be able to dynamically initialize the
pool and react to it if necessary.

Also added an API for changing themes, which, may or may not work?
We'll have to see.  I'll need to probably test it out on the docs site